requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==1.24.3
pillow==10.0.1
rasterio==1.3.9
earthengine-api==0.1.374
python-dotenv==1.0.0
pydantic==2.4.2
scipy==1.11.3
requests==2.31.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-storage==2.10.0
